time7_gateway/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
supabase